create authentication

+++ backend/src/routes/memories.ts
@@ -5,14 +5,22 @@ import { prisma } from '../lib/prisma'
 
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get('/memories', async () => {
+  app.addHook('preHandler', async(request) => {
+    await request.jwtVerify()
+  })
+
+  app.get('/memories', async (request) => {
+    
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: request.user.sub
+      },
       orderBy: {
         createdAt: 'asc',
       }
     })
 
-    return memories.map(memory => {
+    return memories.map((memory: { id: any; coverUrl: any; content: string }) => {
       return {
         id: memory.id,
         coverUrl: memory.coverUrl,
@@ -22,7 +30,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
     })
   })
 
-  app.get('/memories/:id', async (request) => {
+  app.get('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -34,6 +42,10 @@ export async function memoriesRoutes(app: FastifyInstance) {
         id,
       }
     })
+
+    if(!memory.isPublic && memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
 
     return memory
   })
@@ -52,7 +64,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: '2c56b28b-2b87-4207-90bc-50f0c085b454',
+        userId: request.user.sub,
       }
     })
 
@@ -60,7 +72,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
     
   })
 
-  app.put('/memories/:id', async (request) => {
+  app.put('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -73,9 +85,19 @@ export async function memoriesRoutes(app: FastifyInstance) {
       isPublic: z.coerce.boolean().default(false)
     })
 
-    const { content , coverUrl,  isPublic} = bodySchema.parse(request.body)   
+    const { content , coverUrl,  isPublic} = bodySchema.parse(request.body)
     
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      }
+    })
+
+    if(memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
+    
+     memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -88,12 +110,22 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
     return memory
   })
-  app.delete('/memories/:id', async (request) => {
+  app.delete('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
 
     const { id } = paramsSchema.parse(request.params)
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      }
+    })
+
+    if(memory.userId !== request.user.sub) {
+      return reply.status(401).send()
+    }
 
     await prisma.memory.delete({
       where: {
